requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-filter==22.1
django-phone-field==1.8.1
django-restful-admin==1.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ghasedakpack==0.1.13
idna==3.4
Pillow==9.4.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
